Remove span-helper date properties from PR #395
Drop the out-of-scope direct date-matching property module while keeping
production matching and the existing public wrapping properties unchanged.

Keep the integration-only sequence strategy with its remaining consumers.
Track focused direct matching coverage separately in #400.
